+++ R/explain.r
@@ -5,6 +5,12 @@ show_query.tbl_sql <- function(x, ...) {
 }
 
 #' @export
+show_query.tbl_lazy <- function(x, ...) {
+  qry <- sql_build(x, con = x$src, ...)
+  sql_render(qry, con = x$src, ...)
+}
+
+#' @export
 explain.tbl_sql <- function(x, ...) {
   force(x)
   show_query(x)